perf(chronos): replace DatePipe with `formateDate`
We incorrectly created an instance of `DatePipe` to transform
the active timeline segment's timestamp. This was because Angular
doesn't allow using pipes in host bindings; so this was a way to
circumvent that limitation.

Since we import `CommonModule`, which already provides the `date` pipe,
we basically have duplicate instances.

Now we make use of the `formatDate` function provided `@angular/common`
package to format the segment's timestamp.