fix(chronos): inaccurate records layout positioning
Previously, we used `config.offsetStart` and `config.offsetEnd`
to account for the timeline's layout geometry offsets.
Since the `top` and `bottom` css properties were equal to those offsets
we didn't notice any deviations in terms of positioning accuracy.

Further `config.offsetStart` and `config.offsetEnd` are only considered
for usage within the `InteractionManager` class, which returns geometry
metadata without making assumptions about how we how we lay out items.

This commit temporarily adds a `layoutStartOffset` and
`layoutStartOffset` property to the `NgxChronosTimeline` component
to use those for determining the start and end position of the layout
container (the container were we place 'points in time' coordinates and
define layout segments).