feat(chronos: improve accuracy of timeline geometry calculations
* docs(chronos): remove backticks causing unreadable jsdoc

* refactor(chronos): return exact clamped positions

We previously rounded values returned by the `InteractionManager.clamp`
method to the nearest two decimal places. This resulted in inexact
position values in other methods within the InteractionManager class
that use the `clamp()` method.

This commit removes the rounding of the returned value and now returns
the unmodified, exact position with variable decimal places.

* feat(chronos): add enum to determine direction of travel

* refactor(chronos): `normalizeWheelDistance` to `getDirectionOfTravelFromWheel`

This commit better reflects the return value of the previously named
`normalizeWheelDistance` position utility.

* test(chronos): fix failed wheel event handler test

* feat(chronos): add getter that returns the total offset

* feat(chronos): add getter that returns the max space available

* feat(chronos): add helper to calculate relative fractions

* style(chronos): explicitly set return types in public  methods

* feat(chronos): expose computed position percentage in `InteractionManager`

* feat(chronos): add subject to listen to position changes in `InteractionManager`

* feat(chronos): add generic helper to move by distance units

* feat(chronos): derive 'aria-valuenow' directly from InteractionManager

* fix(chronos): inaccurate records layout positioning

Previously, we used `config.offsetStart` and `config.offsetEnd`
to account for the timeline's layout geometry offsets.
Since the `top` and `bottom` css properties were equal to those offsets
we didn't notice any deviations in terms of positioning accuracy.

Further `config.offsetStart` and `config.offsetEnd` are only considered
for usage within the `InteractionManager` class, which returns geometry
metadata without making assumptions about how we how we lay out items.

This commit temporarily adds a `layoutStartOffset` and
`layoutStartOffset` property to the `NgxChronosTimeline` component
to use those for determining the start and end position of the layout
container (the container were we place 'points in time' coordinates and
define layout segments).