fix(chronos): inaccurate segment delimiter layout position
Previously, the computed layout for positioning segments within the
timeline used the total line segment to calculate the final segment
positions.

Additionally, each segment was delimited by a point that had
a fixed length specified in the css styles (in this case `6px`,
including padding), yet we didn't account for that in the layout
calculations.

This commit adds a few configuration options to the TimelineLayoutConfig
interface, such as startOffset, endOffset, and segmentDelimiterWidth,
to accurately calculate each segment's delimiter's final position within
the timeline.

Closes #18